Serving the previous animation code via HTTP
Previously application was CLI based and had too many flags that needed to be configured and also rerun everytime a new output was required. This is solved by migrating the application to be served as service.

For example - Right now we can just call the REST api for get the html code of an animation and display the html directly.